fix: unchanged files get this release's graph; calls leave the definition, not the module
A store swept under an older extractor has every unchanged file's symbols
and edges written again on its next full pass, without re-embedding, so an
upgrade gets method owners and receiver types without an edit to each file.
A function named like its file (fn brief in brief.rs) now owns its calls
rather than the file's module symbol, which put every such caller at line 1.
.semlithignore's skips are counted with the run's skipped reasons.

+++ src/mcp.rs
@@ -1048,11 +1048,11 @@ fn call_tool(
                         const FEW: usize = 5;
                         let (tail_langs, tail_files) = coverage
                             .iter()
-                            .filter(|r| r.definitions > 0 && (r.files as usize) < FEW)
-                            .fold((0usize, 0usize), |(l, f), r| (l + 1, f + r.files as usize));
+                            .filter(|r| r.definitions > 0 && r.files < FEW)
+                            .fold((0usize, 0usize), |(l, f), r| (l + 1, f + r.files));
                         for row in coverage
                             .iter()
-                            .filter(|r| r.definitions > 0 && r.files as usize >= FEW)
+                            .filter(|r| r.definitions > 0 && r.files >= FEW)
                         {
                             body.push_str(&format!(
                                 "\n  {}: {} files ({} unparsed), {} definitions, call edges \
